Stopwatch
Bare bone app written in Swift 1.2 for educational purposes. App is
used to start the stopwatch, reset the timer, and to pause the timer.

Known bug:
If the user presses on the play button more than once, the timer will
increase at twice the rate (instead of increment of 1 per second, it’ll
increment by  “play presses” per second)